+++ R/emulate.R
@@ -101,28 +101,57 @@ emulate.telemetry <- function(object,CTMM,fast=FALSE,...)
 }
 
 
+# remove zero parameters
+ctmm.reduce <- function(CTMM)
+{
+  # parameters to extract & debias
+  NAMES <- dimnames(CTMM$COV)[[1]]
+
+  # delete zero parameters
+  if(length(NAMES))
+  {
+    P <- get.parameters(CTMM,NAMES)
+    ZERO <- !P & (NAMES %in% POSITIVE.PARAMETERS)
+    names(ZERO) <- NAMES
+
+    if("minor" %in% NAMES && ZERO['major']) { ZERO['minor'] <- ZERO['angle'] <- TRUE }
+    P <- "tau velocity"; if(P %in% NAMES && ZERO[P]) { CTMM$tau <- CTMM$tau[1] }
+    P <- "tau position"; if(P %in% NAMES && ZERO[P]) { CTMM$tau <- NULL }
+    P <- "tau"; if(P %in% NAMES && ZERO[P]) { CTMM$tau <- NULL }
+
+    if(any(ZERO)) { CTMM$features <- CTMM$features[ CTMM$features %nin% NAMES[ZERO] ] }
+    NAMES <- NAMES[!ZERO]
+    CTMM$COV <- CTMM$COV[!ZERO,!ZERO]
+  }
+
+  return(CTMM)
+}
+
+
 #####################################
 # multi-estimator parametric bootstrap
 # + concurrent double-bootstrap AICc
-ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores=1,trace=TRUE,...)
+ctmm.boot <- function(data,CTMM,method=CTMM$method,iterate=FALSE,robust=FALSE,error=0.01,cores=1,trace=TRUE,...)
 {
   if("COV" %nin% names(CTMM)) { stop("CTMM needs to be output from ctmm.select or ctmm.fit.") }
 
-  method <- match.arg(method,c('ML','HREML','pREML','pHREML','REML'),several.ok=TRUE)
-  cores <- resolveCores(cores,fast=FALSE)
-
-  # toss location information for simulations unconditioned
-  FRAME <- data
-  FRAME[,CTMM$axes] <- NULL
+  CTMM <- ctmm.reduce(CTMM) # remove zero parameters
+  if(!length(CTMM$COV)) { return(CTMM) } # nothing to do
 
   # parameters to extract & debias
   NAMES <- dimnames(CTMM$COV)[[1]]
-  DIM <- length(NAMES) * length(method)
 
   # positive variables (potential)
   POS <- POSITIVE.PARAMETERS
   # included variables that are positive
   POS <- POS[POS %in% NAMES]
+
+  method <- match.arg(method,c('ML','HREML','pREML','pHREML','REML'),several.ok=FALSE)
+  cores <- resolveCores(cores,fast=FALSE)
+
+  # toss location information for simulations unconditioned
+  FRAME <- data
+  FRAME[,CTMM$axes] <- NULL
 
   Transform <- function(p,inverse=FALSE)
   {
@@ -158,25 +187,16 @@ ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores
   precompute <- FALSE
   Replicate <- function(i=0,DATA=simulate(CTMM,data=FRAME,precompute=precompute),...)
   {
-    # apply all estimators, recycling previous fit as first guess
-    FIT <- list(CTMM=CTMM) # first guess (to be removed)
-    for(m in method) { FIT[[length(FIT)+1]] <- ctmm.fit(DATA,FIT[[length(FIT)]],method=m,COV=FALSE,...) }
-    FIT <- FIT[-1] # remove first (fake) guess
-
-    # extract parameters
-    FIT <- sapply(FIT,function(f){ get.parameters(f,NAMES) }) # [parameter,method]
-
-    rownames(FIT) <- NAMES
-    colnames(FIT) <- method
-
+    FIT <- ctmm.fit(DATA,CTMM,method=method,COV=FALSE,...)
+    FIT <- get.parameters(FIT,NAMES)
+    names(FIT) <- NAMES
     FIT <- Transform(FIT)
-
     return(FIT)
   }
 
   # add batch of calculations to intermediate results
   # rolling mean and variance
-  rolling_update <- function(ADD) # [par*method,run]
+  rolling_update <- function(ADD) # [par,run]
   {
     N <<- N + ncol(ADD) # update sample size
     SAMPLE <<- cbind(SAMPLE,ADD) # update sample (transformed)
@@ -192,7 +212,7 @@ ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores
 
       if(any(abs(AVE)==Inf)) { stop('Some scale parameter estimates on boundary. Requires robust=TRUE.') }
     }
-    else if(N>=DIM+1)
+    else if(N>=length(NAMES)+1)
     {
       STUFF <- rcov(SAMPLE)
       AVE <<- STUFF$median
@@ -202,24 +222,33 @@ ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores
     }
   }
 
-  W <- array(1/length(method),c(length(NAMES),length(method))) # weights
-  J <- array(1,c(length(method),1)) # de-projection matrix
+  # estimate truth from original estimate, correcting for estimated bias
+  estimator <- function(STEP=1)
+  {
+    P <- EST - BIAS
+    names(P) <- NAMES
+    # fixes boundary crossings, etc.
+    P <- Transform(P,inverse=TRUE)
+    P <- clean.parameters(P)
+    P <- Transform(P)
+
+    return(P)
+  }
 
   ### outer loop: designing the weight matrix until estimate converges
-  EST <- Replicate(0,DATA=data) # original estimates that we are working with to optimally weight (transformed)
-  dim(EST) <- DIM # (NAMES*methods)
-  P0 <- P1 <-  Transform(get.parameters(CTMM,NAMES)) # null model for parametric bootstrap
+  P0 <- P1 <- EST <- Transform(get.parameters(CTMM,NAMES)) # null model for parametric bootstrap
   if(trace>1) { print(Transform(P0,inverse=TRUE)) }
   ERROR <- Inf -> ERROR.OLD
+  BEST <- list()
   tol <- error*sqrt(length(NAMES)) # eigen-value tolerance for pseudo-inverse
-  MAX <- max(1/error^2,DIM+1) # maximum sample size
+  MAX <- max(1/error^2,length(NAMES)+1) # maximum sample size
   if(trace) { pb <- utils::txtProgressBar(style=3) }
   while(ERROR>error)
   {
     # initialize results
     SAMPLE <- NULL # sampling distribution sample (transformed)
-    AVE <- array(0,DIM) # average of EST
-    COV <-  array(0,c(DIM,DIM)) # covariation of EST
+    AVE <- array(0,length(NAMES)) # average of EST
+    COV <-  array(0,c(length(NAMES),length(NAMES))) # covariation of EST
     if(!robust) # more efficient computation with moment accumulation
     {
       S1 <- AVE
@@ -230,8 +259,8 @@ ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores
 
     # INITIAL simulation precomputes Kalman filter matrix
     precompute <- TRUE
-    ADD <- Replicate() # [par,method] (transformed)
-    dim(ADD) <- c(DIM,1) # [par*method,1]
+    ADD <- Replicate() # [par] (transformed)
+    dim(ADD) <- c(length(NAMES),1) # [par,1]
     rolling_update(ADD)
 
     # inner loop: adding to ensemble until average converges
@@ -241,78 +270,69 @@ ctmm.boot <- function(data,CTMM,method=CTMM$method,robust=FALSE,error=0.01,cores
       # calculate burst of cores estimates
       ADD <- plapply(1:cores,Replicate,cores=cores,fast=FALSE)
       ADD <- simplify2array(ADD) # [par,method,run]
-      dim(ADD) <- c(DIM,cores) # [par*method,run]
+      dim(ADD) <- c(length(NAMES),cores) # [par*method,run]
       rolling_update(ADD)
 
       # alternative test if bias is relatively well resolved
-      if(N>=DIM+1)
+      if(N>=length(NAMES)+1)
       {
-        BIAS <- AVE - EST
-        ERROR <- sqrt(sum(BIAS^2/diag(COV))/N/DIM)
+        BIAS <- AVE - P0 # fixed for multiple iterations
+        # ERROR <- sqrt(sum(BIAS^2/diag(COV))/length(NAMES)/N)
+        ERROR <- sqrt(c(BIAS %*% PDsolve(COV) %*% BIAS)/length(NAMES)/N)
         if(is.nan(ERROR)) { ERROR <- Inf }
       }
 
       if(trace) { utils::setTxtProgressBar(pb,clamp(max(N/MAX,(error/ERROR)^2))) }
     } # END INNER LOOP
 
-    ## recalculate weighted estimator
-    ## !!! add optimal method to use cross-parameter estimate covariance
-    JACOB <- array(0,c(length(NAMES),DIM))
-    for(i in 1:length(NAMES))
-    {
-      SUB <- array(FALSE,c(length(NAMES),length(method)))
-      SUB[i,] <- TRUE
-
-      H <- PDsolve(COV[SUB,SUB],pseudo=TRUE,tol=tol)
-      VAR <- PDsolve(t(J) %*% H %*% J)
-
-      W[i,] <- VAR %*% t(J) %*% H
-      P1[i] <- c(W[i,] %*% (EST[SUB]-BIAS[SUB]))
-
-      JACOB[i,] <- array(W*SUB,DIM)
-    }
-    # covariance of weighted estimator
-    COV.W <- JACOB %*% COV %*% t(JACOB)
-
     ## recalculate error & store best model
-    ERROR <- P1 - P0 # how much does best estimate change?
-    ERROR <- sqrt(abs(c(ERROR %*% PDsolve(COV.W) %*% ERROR))/length(NAMES)/2) # 2-point comparison correction
+    ERROR <- AVE - EST # how far off were we?
+    ERROR <- sqrt(abs(c(ERROR %*% PDsolve(COV) %*% ERROR))/length(NAMES))
     if(is.nan(ERROR)) { ERROR <- Inf }
 
     # check to make sure relative error is decreasing
-    if(ERROR>ERROR.OLD)
+    if(ERROR<ERROR.OLD) # made progress
     {
-      # maybe allow this once or twice with a count???
-      if(trace>1) { warning("Iterative convergence ceased at error=",ERROR.OLD," over tolerance of ",error)  }
-      break
+      ERROR -> ERROR.OLD
+
+      BEST$COV <- COV
+      BEST$BIAS <- BIAS
+
+      # final iteration --- calculate final parameter estimate COV
+      if(ERROR<=error || !iterate)
+      {
+        ## calculate COV (somewhat robustly in untransformed basis)
+        # calculate all weighted esitmates in sample
+        SAMPLE <- SAMPLE - c(BIAS) # debiased estimates
+        dim(SAMPLE) <- c(length(NAMES),N)
+        rownames(SAMPLE) <- NAMES
+        SAMPLE <- Transform(SAMPLE,inverse=TRUE)
+
+        # robust covariance estimate
+        # if(!robust) { COV <- cov(t(SAMPLE)) } else { COV <- rcov(SAMPLE) }
+        COV <- stats::cov(t(SAMPLE))
+        dimnames(COV) <- list(NAMES,NAMES)
+        CTMM$COV <- COV
+      }
     }
-    ERROR -> ERROR.OLD
+    else # back up to previous and quit
+    {
+      # go back to previous (best) result
+      BEST$COV -> COV
+      BEST$BIAS -> BIAS
+    }
 
-    # order of current error, which we do not know
-    ERROR <- ERROR^2
-    names(P1) <- NAMES
+    ## calculate best estimate
+    P1 <- estimator()
+
     P <- Transform(P1,inverse=TRUE)
-    P <- clean.parameters(P) # fixes boundary crossings, etc.
-    if(trace>1) { print(P) }
     CTMM <- set.parameters(CTMM,P)
+    if(trace) { message("\nRelative error = ",ERROR) }
     P1 -> P0
+
+    if(!iterate || ERROR>ERROR.OLD) { break }
+    if(trace>1) { print(P) }
   } # END OUTER LOOP
-
-  ## calculate COV (somewhat robustly in untransformed basis)
-  # calculate all weighted esitmates in sample
-  SAMPLE <- SAMPLE - c(BIAS) # debiased estimates
-  dim(SAMPLE) <- c(length(NAMES),length(method),N)
-  SAMPLE <- sapply(1:length(NAMES),function(p) W[p,] %*% array(SAMPLE[p,,],c(length(method),N)) ) # R arrays are insane
-  dim(SAMPLE) <- c(N,length(NAMES)) # R arrays are insane, part 2
-  SAMPLE <- t(SAMPLE)
-  rownames(SAMPLE) <- NAMES
-  SAMPLE <- Transform(SAMPLE,inverse=TRUE)
-
-  # robust covariance estimate
-  # if(!robust) { COV <- cov(t(SAMPLE)) } else { COV <- rcov(SAMPLE) }
-  COV <- stats::cov(t(SAMPLE))
-  dimnames(COV) <- list(NAMES,NAMES)
-  CTMM$COV <- COV
 
   # fix COV[mean] with a verbose likelihood evaluation
   CTMM <- ctmm.loglike(data,CTMM,profile=FALSE,verbose=TRUE)